requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
cryptography==45.0.5
DAWG2-Python==0.9.0
joblib==1.5.1
nltk==3.9.1
numpy==2.3.2
pdfminer.six==20250506
plum-dispatch==1.7.4
pycparser==2.22
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
PyPDF2==3.0.1
regex==2025.7.34
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
spire-doc==13.6.4
threadpoolctl==3.6.0
tqdm==4.67.1